refactor(sse-chat): split monolithic main.rs into modules
