+++ examples/sse-chat/src/config.rs
@@ -1,0 +1,10 @@
+use modo::AppConfig;
+use modo_db::DatabaseConfig;
+use serde::Deserialize;
+
+#[derive(Default, Deserialize)]
+pub(crate) struct Config {
+    #[serde(flatten)]
+    pub(crate) core: AppConfig,
+    pub(crate) database: DatabaseConfig,
+}
