+++ examples/sse-chat/src/entity.rs
@@ -1,0 +1,10 @@
+#[modo_db::entity(table = "messages")]
+#[entity(timestamps)]
+pub struct Message {
+    #[entity(primary_key, auto = "ulid")]
+    pub id: String,
+    #[entity(indexed)]
+    pub room: String,
+    pub username: String,
+    pub text: String,
+}
